tmux: Add a prefix option overruling shortcut if defined
It was not possible to set an arbitrary tmux prefix, ctrl was hardcoded
in the module.
To avoid breaking existing configs I implemented a new option which can
conveniently use the tmux terminology but defaults to null and doesn't
affect behavior when set to null.
I did not completely replicate the behavior for the shortcut option,
i.e., I do not bind "b" to send-prefix but stick to the default of the
prefix binding sending prefix (C-b C-b instead of C-b b) and I do not
bind repetition of the prefix (C-b C-b) to `last-window`, both of these
bring the option closer to the default tmux configuration.

Fixes #1237